:ref: moved tests in same module as tested classes
