+++ service/security/authorization/api/src/test/java/org/eclipse/kapua/service/authorization/exception/KapuaAuthorizationExceptionTest.java
@@ -10,16 +10,13 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.service.authorization.shiro.exception;
+package org.eclipse.kapua.service.authorization.exception;
 
 import org.eclipse.kapua.qa.markers.junit.JUnitTests;
-import org.eclipse.kapua.service.authorization.exception.KapuaAuthorizationErrorCodes;
-import org.eclipse.kapua.service.authorization.exception.KapuaAuthorizationException;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
-
 
 @Category(JUnitTests.class)
 public class KapuaAuthorizationExceptionTest {
